% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignraster.R
\name{alignraster}
\alias{alignraster}
\title{Combines region-specific group-level UD rasters into a single raster.}
\usage{
alignraster(
  folderroots =
    c("/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/H",
    "/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/L",
    "/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/M"),
  foldernames = c("H", "L", "M"),
  pattern = ".asc",
  scalefolder = "Scaled",
  scaledweightedname = "All_Rasters_Scaled",
  savefolder =
    "/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/Aligned",
  format = "ascii",
  datatype = "FLT4S",
  bylayer = TRUE,
  overwrite = TRUE,
  returnObj = FALSE
)
}
\arguments{
\item{folderroots}{Character vector of locations of folder roots output by dBBMMhomeRange. Function expects CRS.Rds file and a subfolder with the scaled raster.}

\item{foldernames}{Character vector names of folders corresponding to files in folderroots, i.e. the names of the objects, arrays, regions, etc.}

\item{pattern}{For input rasters from scaleraster. Default ".asc".}

\item{scalefolder}{For input rasters from scaleraster. Default "Scaled".}

\item{scaledweightedname}{For input rasters from scaleraster. Default "All_Rasters_Scaled".}

\item{savefolder}{Single character entry, no trailing slash.}

\item{format}{Character. Output file type. ascii. (Mo: i think we need to list options to choose from. what is default?).}

\item{datatype}{Character. Data type for writing values to disk. (Mo: should we mention FLT4S? Should we give choices? what is default?).}

\item{bylayer}{Default TRUE.}

\item{overwrite}{Default TRUE.}

\item{returnObj}{Logical. Return the scaled object to the parent environment? Default FALSE.}
}
\value{
Region-specific group-level UD rasters that share the same spatial extent.
}
\description{
Extends the spatial extent of each area-specific group-level raster to the spatial extent shared by all rasters.
}
\details{
Errors and their origins:
}
\examples{
\donttest{
# Not run
}

}
\author{
Simon Dedman, \email{simondedman@gmail.com}

Maurits van Zinnicq Bergmann, \email{mauritsvzb@gmail.com}
}
