% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBBMM.build.R
\name{dBBMMhomeRange}
\alias{dBBMMhomeRange}
\title{Automates dynamic Brownian bridge construction}
\usage{
dBBMMhomeRange(
  data = NULL,
  ID = NULL,
  Datetime = NULL,
  Lat = NULL,
  Lon = NULL,
  Group = NULL,
  dat.TZ = "US/Eastern",
  proj = sp::CRS("+proj=longlat +datum=WGS84"),
  projectedCRS = "+init=epsg:32617",
  sensor = "VR2W",
  moveLocError = 1,
  timeDiffLong = 2,
  timeDiffUnits = "hours",
  center = TRUE,
  buffpct = 0.3,
  rasterExtent = NULL,
  rasterCRS = sp::CRS("+proj=utm +zone=17 +datum=WGS84"),
  rasterResolution = 50,
  bbdlocationerror = "LocationError",
  bbdext = 3,
  bbdwindowsize = 23,
  writeRasterFormat = "ascii",
  writeRasterExtension = ".asc",
  writeRasterDatatype = "FLT4S",
  absVolumeAreaSaveName = "VolumeArea_AbsoluteScale.csv",
  savedir = tempdir(),
  alerts = TRUE
)
}
\arguments{
\item{data}{Data frame of data needs columns Lat Lon DateTime and optionally an ID and grouping column.}

\item{ID}{Column name of IDs of individuals.}

\item{Datetime}{Name of Datetime column. Must be in POSIXct format.}

\item{Lat}{Name of Lat & Lon columns in data.}

\item{Lon}{Name of Lat & Lon columns in data.}

\item{Group}{Name of grouping column in data. CURRENTLY UNUSED; MAKE USER DO THIS?.}

\item{dat.TZ}{Timezone for as.POSIXct.}

\item{proj}{CRS for move function.}

\item{projectedCRS}{EPSG code for CRS for initial transform of latlon points; corresponds to rasterCRS zone.}

\item{sensor}{Sensor for move function. Single character or vector with length of the number of coordinates. Optional.}

\item{moveLocError}{Location error in metres for move function. Numeric. Either single or a vector of lenth nrow data.}

\item{timeDiffLong}{Threshold length of time in timeDiffUnits designating long breaks in relocations.}

\item{timeDiffUnits}{Units for time difference for move function.}

\item{center}{Centre move object within extent? See spTransform.}

\item{buffpct}{Buffer extent for raster creation, proportion of 1.}

\item{rasterExtent}{If NULL, raster extent calculated from data, buffpct, rasterResolution. Else length 4 vector, c(xmn, xmx, ymn, ymx) decimal latlon degrees. Don't go to 90 for ymax. Doesn't prevent constraint to data limits (in plot anyway), but prevents raster clipping crash.}

\item{rasterCRS}{CRS for raster creation.}

\item{rasterResolution}{Single numeric value to set raster resolution - cell size in metres? 111000: 1 degree lat = 111km.}

\item{bbdlocationerror}{Location.error param in brownian.bridge.dyn. Could use the same as moveLocError?.}

\item{bbdext}{Ext param in brownian.bridge.dyn. Extends bounding box around track. Numeric single (all edges), double (x & y), or 4 (xmin xmax ymin ymax). Default 0.3.}

\item{bbdwindowsize}{window.size param in brownian.bridge.dyn. The size of the moving window along the track. Larger windows provide more stable/accurate estimates of the brownian motion variance but are less well able to capture more frequent changes in behavior. This number has to be odd. A dBBMM is not run if total detections of individual < window size (default 31).}

\item{writeRasterFormat}{ascii.}

\item{writeRasterExtension}{.asc.}

\item{writeRasterDatatype}{FLT4S.}

\item{absVolumeAreaSaveName}{VolumeArea_AbsoluteScale.csv.}

\item{savedir}{Save outputs to a temporary directory (default) else. Change to current directory e.g. "/home/me/folder". Do not use getwd() here.}

\item{alerts}{Audio warning for failures.}
}
\value{
Individual-level utilization distributions, saved as rasters, as well as volume area estimates for 50% and 95% contours, saved in a .csv file

saved to disk.
}
\description{
This function automates dynamic Brownian bridge movement model calculation for utilization distribution (UD) estimation for multiple 
individuals simultaneously, via the brownian.bridge.dyn() {move}. It first removes those individuals for which there is insufficient data 
i.e. number of re-locations is smaller than the window size parameter value (default = 31). Second, based on all remaining data, a universal 
raster is generated where the calculated UDs are plotted into. Finally, remaining individuals are looped through to construct individual-level 
movement models (on an absolute scale). See www.GitHub.com/SimonDedman/dBBMMhomeRange for issues, feedback, and development suggestions.
}
\details{
Errors and their origins:
}
\examples{
\donttest{
# Not run
}

}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
