% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBBMM.plot.R
\name{dBBMMplot}
\alias{dBBMMplot}
\title{Succinct title 8 words max}
\usage{
dBBMMplot(
  x = paste0("Scaled/All_Rasters_Scaled_LatLon.asc"),
  trim = TRUE,
  myLocation = NULL,
  googlemap = FALSE,
  gmapsAPI = NULL,
  expandfactor = 1.6,
  mapzoom = 5,
  mapsource = "google",
  maptype = "satellite",
  contour1colour = "red",
  contour2colour = "orange",
  plottitle = "Aggregated 95\% and 50\% UD contours",
  plotsubtitle = "Scaled contours. n = 13",
  legendtitle = "Percent UD Contours",
  plotcaption = paste0("dBBMM_HomeRange, ", lubridate::today()),
  axisxlabel = "Longitude",
  axisylabel = "Latitude",
  legend.position = c(0.16, 0.92),
  filesavename = paste0(lubridate::today(), "_dBBMM-contours.png"),
  savedir = tempdir()
)
}
\arguments{
\item{x}{Path to scaled data.}

\item{trim}{Remove NA & 0 values and crop to remaining date extents? Default TRUE.}

\item{myLocation}{Location for extents, format c(xmin, ymin, xmax, ymax). Default NULL, extents autocreated from data.}

\item{googlemap}{If pulling basemap from Google maps, this sets expansion factors since Google Maps tiling zoom setup doesn't align to myLocation extents.}

\item{gmapsAPI}{Enter your google maps API here, quoted character string.}

\item{expandfactor}{Extents expansion factor for basemap. 1.3 to 1.5 are the same zoom as 1. 1.6 is a big leap up in zoom (out). 1.9 & maybe 1.7 or 1.8 is another step out. Ignored if not using Google Maps.}

\item{mapzoom}{Google: 3 (continent) - 21 (building). stamen: 0-18.}

\item{mapsource}{Source for ggmap::get_map; uses Stamen as fallback if no Goole Maps API present.}

\item{maptype}{Type of map for ggmap::get_map.}

\item{contour1colour}{Colour for contour 1, typically 95pct.}

\item{contour2colour}{Colour for contour 2, typically 50pct.}

\item{plottitle}{Can use the term 'home range' when an animal can be detected wherever it goes
i.e. using GPS, satellite or acoustic telemetry whereby it is known that acoustic receivers cover 
the entire home range of the study species. This term is problematic when applied to a passive 
acoustic telemetry setting where an array of non-overlapping receivers are used to assess local 
space use patterns i.e. the home range is bigger than the coverage by the acoustic array; put in Details.}

\item{plotsubtitle}{Plot subtitle.}

\item{legendtitle}{Legend title.}

\item{plotcaption}{Plot caption.}

\item{axisxlabel}{Longitude.}

\item{axisylabel}{Latitude.}

\item{legend.position}{Vector of 2, format c(1,2), from L to R, pct dist from Bot to Top, values 0 to 1.}

\item{filesavename}{File savename.}

\item{savedir}{Save outputs to a temporary directory (default) else. Change to current directory e.g. "/home/me/folder". Do not use getwd() here. No terminal slash.}
}
\description{
Description paragraph: Automates delta log normal boosted regression trees abundance prediction.
Loops through all permutations of parameters provided (learning
rate, tree complexity, bag fraction), chooses the best, then simplifies it.
Generates line, dot and bar plots, and outputs these and the predictions
and a report of all variables used, statistics for tests, variable
interactions, predictors used and dropped, etc. If selected, generates
predicted abundance maps, and Unrepresentativeness surfaces.
See www.GitHub.com/SimonDedman/gbm.auto for issues, feedback, and development
suggestions. See SimonDedman.com for links to walkthrough paper, and papers
and thesis published using this package.
}
