% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleraster.R
\name{scaleraster}
\alias{scaleraster}
\title{Scales individual utilization distribution rasters and volume area estimates}
\usage{
scaleraster(
  path = NULL,
  pathsubsets = NULL,
  pattern = ".asc",
  weighting = w,
  format = "ascii",
  datatype = "FLT4S",
  bylayer = TRUE,
  overwrite = TRUE,
  scalefolder = "Scaled",
  summedname = "All_Rasters_Summed",
  scaledname = "All_Rasters_Scaled",
  crsloc = NULL,
  returnObj = FALSE
)
}
\arguments{
\item{path}{No terminal slash.}

\item{pathsubsets}{Location of parent folder that contains ALL files created by dBBMM.build. No terminal slash.}

\item{pattern}{Default ".asc".}

\item{weighting}{Weighting to divide scaled group-level raster by, for unbalanced arrays. Scaled_Weighted rasters, and its volume areas csv, will have weightings applied, but NOT the summed raster.  MO NEED TO EDIT THIS}

\item{format}{Default "ascii".}

\item{datatype}{Default "FLT4S".}

\item{bylayer}{Default TRUE.}

\item{overwrite}{Default TRUE.}

\item{scalefolder}{Default "Scaled".}

\item{summedname}{Default "All_Rasters_Summed".}

\item{scaledname}{Default "All_Rasters_Scaled".}

\item{crsloc}{Location of saved CRS Rds file from dBBMM.build.R. Should be same as path.}

\item{returnObj}{Logical. Return the scaled object to the parent environment? Default FALSE.}
}
\value{
scaled individual-level and group-level utilization distributions saved as rasters. Additionally, scaled 50% and 95% 
contour volume area estimates for individuals and the group, saved in .csv format.
}
\description{
Scales individual-level utilization distribution (UD) rasters from 0 to 1 to facilitate interpretation as relative 
intensity of utilization (as opposed to absolute). Subsequently, scaled individual-level rasters are aggregated to 
create a single group-level UD raster. See www.GitHub.com/SimonDedman/dBBMMhomeRange for issues, feedback, and 
development suggestions.
}
\details{
# Step 1 - Scale rasters
Individual-level UD rasters are scaled from 0 to 1 by dividing each raster by the maximum probability density value
occurring within the raster set.

# Step 2 - Aggregate into a group-level raster
Scaled individual-level rasters are summed to create a single group-level UD raster. 


# Step 3 - Re-scale from 0 to 1
The group-level raster is divided by its own maximum value.

# Step 4 - Weight raster (optional)
The scaled group-level UD raster is divided by the specified weighting factor(s). Note that this is only useful if you 
want to account for an unbalanced receiver array and have split up the study site and receivers in clusters, and have 
run the dBBMMhomeRange() for each cluster data set separately. See van Zinnicq Bergmann et al. 2022 
(https://doi.org/10.1016/j.biocon.2022.109469) for example. If not applicable, choose a value of "1".

# Step 5 - Standardize raster 
Standardize the potentially weighted and scaled group-level UD raster so that its values sum to 1.  (creates UDScaled, but is in .UD extension)

# Step 6 - Change crs
Change its crs to latlon for plotting and calculation purposes (DO WE WANT TO INCLUDE THIS? CURRENTLY DOES NOTHING)

# Step 7 - Estimate 50 and 95pct contour volume areas
For each scaled individual-level UD raster, estimate 50 and 95pct contour volume areas, as well as their mean and standard
deviation. Additionally, the 50 and 95pct volume area is estimated for the group-level UD raster.

Errors and their origins.
}
\examples{
\donttest{
# Not run
}

}
\author{
Simon Dedman, \email{simondedman@gmail.com}

Maurits van Zinnicq Bergmann, \email{mauritsvzb@gmail.com}
}
