% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitchraster.R
\name{stitchraster}
\alias{stitchraster}
\title{Stitch together movegroup data individuals core and general use areas}
\usage{
stitchraster(
  data = NULL,
  ID = NULL,
  absVolumeAreaSaveName = "VolumeArea_AbsoluteScale.csv",
  savedir = tempdir()
)
}
\arguments{
\item{data}{Data frame object containing the data. Requires columns Lat Lon DateTime ID and
potentially a grouping column (not currently implemented, email to request). Column names
specified in later parameters.}

\item{ID}{Name of animal tag ID column in data. "Character".}

\item{absVolumeAreaSaveName}{File name plus extension where UD estimates are saved. Default
"VolumeArea_AbsoluteScale.csv".}

\item{savedir}{Save outputs to a temporary directory (default) else change
to desired directory e.g. "/home/me/folder". Do not use getwd() for this.
Do NOT include terminal slash. Directory must exist. Default tempdir().}
}
\value{
Calculated volume area estimates for 50 and 95pct contours csv.
}
\description{
If over-large datasets cause RAM crashes for movegroup, one can run batches
of individuals in movegroup then join the individual saved area.ct csv files.
}
\details{
See www.GitHub.com/SimonDedman/movegroup for issues, feedback, and
development suggestions. Install 'move' development version with:
remotes::install_git('https://gitlab.com/bartk/move.git')

Parameters values should match those used in movegroup.
}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
