% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightraster.R
\name{weightraster}
\alias{weightraster}
\title{Succinct title 8 words max}
\usage{
weightraster(
  locations = c("array1", "array2"),
  weightings = c(1, 2),
  saveloc = "location"
)
}
\arguments{
\item{locations}{No terminal slash.}

\item{weightings}{No terminal slash.}

\item{saveloc}{No terminal slash.}
}
\description{
Description paragraph: Automates delta log normal boosted regression trees abundance prediction.
Loops through all permutations of parameters provided (learning
rate, tree complexity, bag fraction), chooses the best, then simplifies it.
Generates line, dot and bar plots, and outputs these and the predictions
and a report of all variables used, statistics for tests, variable
interactions, predictors used and dropped, etc. If selected, generates
predicted abundance maps, and Unrepresentativeness surfaces.
See www.GitHub.com/SimonDedman/gbm.auto for issues, feedback, and development
suggestions. See SimonDedman.com for links to walkthrough paper, and papers
and thesis published using this package.
}
