% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignraster.R
\name{alignraster}
\alias{alignraster}
\title{Automated Boosted Regression Tree modelling and mapping suite}
\usage{
alignraster(
 
    folderroots = c("/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/H",
    "/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/L",
    "/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/M"),
  foldernames = c("H", "L", "M"),
  pattern = ".asc",
  scalefolder = "Scaled",
  scaledname = "All_Rasters_Scaled",
 
    savefolder = "/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/Aligned",
  format = "ascii",
  datatype = "FLT4S",
  bylayer = TRUE,
  overwrite = TRUE,
  returnObj = FALSE
)
}
\arguments{
\item{folderroots}{Character vector of locations of folder roots output by dBBMMhomeRange. Function expects CRS.Rds file and a subfolder with the scaled raster.}

\item{foldernames}{Character vector names of folders corresponding to files in folderroots, i.e. the names of the objects, arrays, regions, etc.}

\item{pattern}{For input rasters from scaleraster. Default ".asc".}

\item{scalefolder}{For input rasters from scaleraster. Default "Scaled".}

\item{scaledname}{For input rasters from scaleraster. Default "All_Rasters_Scaled".}

\item{savefolder}{Single character entry, no trailing slash.}

\item{format}{Default "ascii".}

\item{datatype}{Default "FLT4S".}

\item{bylayer}{Default TRUE.}

\item{overwrite}{Default TRUE.}

\item{returnObj}{Logical. Return the scaled object to the parent environment? Default FALSE.}
}
\value{
Line, dot and bar plots, a report of all variables used, statistics
for tests, variable interactions, predictors used and dropped, etc. If
selected generates predicted abundance maps, and Unrepresentativeness surface
}
\description{
Automates delta log normal boosted regression trees abundance prediction.
Loops through all permutations of parameters provided (learning
rate, tree complexity, bag fraction), chooses the best, then simplifies it.
Generates line, dot and bar plots, and outputs these and the predictions
and a report of all variables used, statistics for tests, variable
interactions, predictors used and dropped, etc. If selected, generates
predicted abundance maps, and Unrepresentativeness surfaces.
See www.GitHub.com/SimonDedman/gbm.auto for issues, feedback, and development
suggestions. See SimonDedman.com for links to walkthrough paper, and papers
and thesis published using this package.
}
\details{
Errors and their origins:
}
\examples{
\donttest{
# Not run
}

}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
