% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBBMM.plot.R
\name{dBBMMplot}
\alias{dBBMMplot}
\title{Plots a group-level utilization distribution}
\usage{
dBBMMplot(
  x = paste0("Scaled/All_Rasters_Scaled_Weighted_UDScaled.asc"),
  crsloc = NULL,
  xlatlon = NULL,
  locationpoints =
    paste0("/home/simon/Documents/Si Work/PostDoc Work/movegroup help/Liberty Boyd/Points in UD contours/",
    "turtlehab_all.csv"),
  pointsincontourssave =
    paste0("/home/simon/Documents/Si Work/PostDoc Work/movegroup help/Liberty Boyd/Points in UD contours/",
    "pointsincontours.csv"),
  trim = TRUE,
  myLocation = NULL,
  googlemap = FALSE,
  gmapsAPI = NULL,
  expandfactor = 1.6,
  mapzoom = 5,
  mapsource = "google",
  maptype = "satellite",
  contour1colour = "red",
  contour2colour = "orange",
  plottitle = "Aggregated 95\% and 50\% UD contours",
  plotsubtitle = "Scaled contours",
  legendtitle = "Percent UD Contours",
  plotcaption = paste0("movegroup, ", lubridate::today()),
  axisxlabel = "Longitude",
  axisylabel = "Latitude",
  legendposition = c(0.11, 0.85),
  fontsize = 12,
  fontfamily = "Times New Roman",
  filesavename = paste0(lubridate::today(), "_dBBMM-contours.png"),
  savedir = tempdir(),
  receiverlats = NULL,
  receiverlons = NULL,
  receivernames = NULL,
  receiverrange = NULL,
  recpointscol = "black",
  recpointsfill = "white",
  recpointsalpha = 0.5,
  recpointssize = 1,
  recpointsshape = 21,
  recbufcol = "grey75",
  recbuffill = "grey",
  recbufalpha = 0.5,
  reclabcol = "black",
  reclabfill = NA,
  reclabnudgex = 0,
  reclabnudgey = -200,
  reclabpad = 0,
  reclabrad = 0.15,
  reclabbord = 0,
  surface = TRUE
)
}
\arguments{
\item{x}{Path to scaled raster generated by movegroup (/scaleraster/alignraster). Likely 
paste0(movegroupSavedir, "Scaled/All_Rasters_Scaled_Weighted_LatLon.asc").}

\item{crsloc}{Location of saved CRS Rds file from movegroup.R. Should be same as path. Likely 
movegroupSavedir.}

\item{xlatlon}{If you want to also return a csv of your original locations labelled with which UD
 contours they fall within, include the location of that here. Try 
 paste0(crsloc, "Scaled/All_Rasters_Scaled_Weighted_LatLon.asc") . Default NULL will not produce 
the csv output.}

\item{locationpoints}{Original input location points of animals, for xlatlon. MUST have columns labelkled "lat" and "lon".}

\item{pointsincontourssave}{Location and name of location in countours csv, including the .csv.}

\item{trim}{Remove NA & 0 values and crop to remaining date extents? Shrinks lots of dead space 
at the edges of the raster. Default TRUE.}

\item{myLocation}{Location for extents, format c(xmin, ymin, xmax, ymax). Default NULL, extents 
autocreated from data.}

\item{googlemap}{If pulling basemap from Google maps, this sets expansion factors since Google 
Maps tiling zoom setup doesn't align to myLocation extents. Default FALSE.}

\item{gmapsAPI}{Enter your google maps API here, quoted character string. Can leave NULL if 
already registered with ggmap::register_google().}

\item{expandfactor}{Extents expansion factor for basemap. 1.3 to 1.5 are the same zoom as 1. 1.6 
is a big leap up in zoom (out). 1.9 & maybe 1.7 or 1.8 is another step out. Ignored if not using 
Google Maps.}

\item{mapzoom}{Highest number = zoomed in. Google: 3 (continent) - 21 (building). stamen: 0-18.}

\item{mapsource}{Source for ggmap::get_map; uses Stamen as fallback if no Google Maps API present.}

\item{maptype}{Type of map for ggmap::get_map.}

\item{contour1colour}{Colour for contour 1, typically 95pct.}

\item{contour2colour}{Colour for contour 2, typically 50pct.}

\item{plottitle}{Can use the term 'home range' when an animal can be detected wherever it goes
i.e. using GPS, satellite or acoustic telemetry whereby it is known that acoustic receivers cover 
the entire home range of the study species. This term is problematic when applied to a passive 
acoustic telemetry setting where an array of non-overlapping receivers are used to assess local 
space use patterns i.e. the home range is bigger than the coverage by the acoustic array; put in 
Details.}

\item{plotsubtitle}{Plot subtitle. If NULL (default) will print "Scaled contours". Can add the n 
of your individuals.}

\item{legendtitle}{Legend title.}

\item{plotcaption}{Plot caption.}

\item{axisxlabel}{Longitude.}

\item{axisylabel}{Latitude.}

\item{legendposition}{Vector of 2, format c(1,2), Percent distance (of middle? of legend box) 
from L to R, percent distance from Bottom to Top, values 0 to 1.}

\item{fontsize}{Font size, default 12.}

\item{fontfamily}{= Font family, default "Times New Roman".}

\item{filesavename}{File savename.}

\item{savedir}{Save outputs to a temporary directory (default) else. Change to current directory 
e.g. "/home/me/folder". Do not use getwd() here. No terminal slash. E.g. 
paste0(movegroupSavedir, "Plot/") .}

\item{receiverlats}{Vector of latitudes for receivers to be plotted.}

\item{receiverlons}{Vector of longitudes for receivers to be plotted.}

\item{receivernames}{Vector of names for receivers to be plotted.}

\item{receiverrange}{Single (will be recycled), or vector of detection ranges in metres for 
receivers to be plotted. If you have a max and a (e.g.) 90 percent detection range, probably use 
max.}

\item{recpointscol}{Colour of receiver centrepoint outlines.}

\item{recpointsfill}{Colour of receiver centrepoint fills.}

\item{recpointsalpha}{Alpha value of receiver centrepoint fills, 0 (invisible) to 1 (fully 
visible).}

\item{recpointssize}{Size of receiver points.}

\item{recpointsshape}{Shape of receiver points, default 21, circle with outline and fill.}

\item{recbufcol}{Colour of the receiver buffer circle outlines.}

\item{recbuffill}{Colour of the receiver buffer circle fills.}

\item{recbufalpha}{Alpha value of receiver buffer fills, 0 (invisible) to 1 (fully visible).}

\item{reclabcol}{Receiver label text colour.}

\item{reclabfill}{Receiver label fill colour, NA for no fill.}

\item{reclabnudgex}{Receiver label offset nudge in X dimension.}

\item{reclabnudgey}{Receiver label offset nudge in Y dimension.}

\item{reclabpad}{Receiver label padding in lines.}

\item{reclabrad}{Receiver label radius in lines.}

\item{reclabbord}{Receiver label border in mm.}

\item{surface}{Plot complete UD surface along with contours.}
}
\value{
Individual-level utilization distributions, saved as rasters, as well as calculated 
volume area estimates for 50 and 95pct contours, saved in a .csv file.
}
\description{
Plots 50 and 95pct contours of a group-level utilization distribution raster on a spatial map 
background. Contains functionality to also visualize geographic locations of individual listening
stations (e.g., acoustic receivers) as well as the entire surface UD.
}
\details{
Errors and their origins:
1. Error in gzfile(file, "rb") : cannot open the connection. In addition: Warning message: In 
gzfile(file, "rb"): cannot open compressed file '/var/folders/dl/gobbledigook/ggmap/index.rds', 
probable reason 'No such file or directory'. Cause: index.rds may not have been created, due to a
problem with ggmap::get_map, likely due to your API key failing silently.
Filename too long - solve with filename = "whatever" but doesn't do anything. Added issue in 
github: https://github.com/dkahle/ggmap/issues/346 .
API key help: https://github.com/dkahle/ggmap/issues/235 .
2. trying to read file: All_Rasters_Scaled_Weighted_UDScaled.asc: Error in CPL_read_gdal(
as.character(x), as.character(options), as.character(driver),: file not found. Check x is correct.

How to get Google map basemaps:

#' @examples
\donttest{
# Not run
}
}
\author{
Simon Dedman, \email{simondedman@gmail.com}

Maurits van Zinnicq Bergmann, \email{mauritsvzb@gmail.com}
}
