% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBBMM.plot.R
\name{dBBMMplot}
\alias{dBBMMplot}
\title{Plots a group-level utilization distribution on a map background}
\usage{
dBBMMplot(
  x = paste0("Scaled/All_Rasters_Scaled_Weighted_UDScaled.asc"),
  crsloc = NULL,
  trim = TRUE,
  myLocation = NULL,
  googlemap = FALSE,
  gmapsAPI = NULL,
  expandfactor = 1.6,
  mapzoom = 5,
  mapsource = "google",
  maptype = "satellite",
  contour1colour = "red",
  contour2colour = "orange",
  plottitle = "Aggregated 95\% and 50\% UD contours",
  plotsubtitle = "Scaled contours. n = 13",
  legendtitle = "Percent UD Contours",
  plotcaption = paste0("dBBMM_HomeRange, ", lubridate::today()),
  axisxlabel = "Longitude",
  axisylabel = "Latitude",
  legendposition = c(0.11, 0.85),
  fontsize = 12,
  fontfamily = "Times New Roman",
  filesavename = paste0(lubridate::today(), "_dBBMM-contours.png"),
  savedir = tempdir(),
  receiverlats = NULL,
  receiverlons = NULL,
  receivernames = NULL,
  receiverrange = NULL,
  recpointscol = "black",
  recpointsfill = "white",
  recpointsalpha = 0.5,
  recpointssize = 1,
  recpointsshape = 21,
  recbufcol = "black",
  recbuffill = "red",
  recbufalpha = 0.5,
  reclabcol = "black",
  reclabfill = NA,
  reclabnudgex = 0,
  reclabnudgey = -200,
  reclabpad = 0,
  reclabrad = 0.15,
  reclabbord = 0,
  surface = TRUE
)
}
\arguments{
\item{x}{Path to scaled raster.}

\item{crsloc}{Location of saved CRS Rds file from dBBMM.build.R. Should be same as path.}

\item{trim}{Remove NA & 0 values and crop to remaining date extents? Default TRUE.}

\item{myLocation}{Location for extents, format c(xmin, ymin, xmax, ymax). Default NULL, extents autocreated from data.}

\item{googlemap}{If pulling basemap from Google maps, this sets expansion factors since Google Maps tiling zoom setup doesn't align to myLocation extents.}

\item{gmapsAPI}{Enter your google maps API here, quoted character string.}

\item{expandfactor}{Extents expansion factor for basemap. 1.3 to 1.5 are the same zoom as 1. 1.6 is a big leap up in zoom (out). 1.9 & maybe 1.7 or 1.8 is another step out. Ignored if not using Google Maps.}

\item{mapzoom}{Google: 3 (continent) - 21 (building). stamen: 0-18.}

\item{mapsource}{Source for ggmap::get_map; uses Stamen as fallback if no Google Maps API present.}

\item{maptype}{Type of map for ggmap::get_map.}

\item{contour1colour}{Colour for contour 1, typically 95pct.}

\item{contour2colour}{Colour for contour 2, typically 50pct.}

\item{plottitle}{Can use the term 'home range' when an animal can be detected wherever it goes
i.e. using GPS, satellite or acoustic telemetry whereby it is known that acoustic receivers cover 
the entire home range of the study species. This term is problematic when applied to a passive 
acoustic telemetry setting where an array of non-overlapping receivers are used to assess local 
space use patterns i.e. the home range is bigger than the coverage by the acoustic array; put in Details.}

\item{plotsubtitle}{Plot subtitle.}

\item{legendtitle}{Legend title.}

\item{plotcaption}{Plot caption.}

\item{axisxlabel}{Longitude.}

\item{axisylabel}{Latitude.}

\item{legendposition}{Vector of 2, format c(1,2), Percent distance (of middle? of legend box) from L to R, percent distance from Bottom to Top, values 0 to 1.}

\item{fontsize}{Font size, default 12.}

\item{fontfamily}{= Font family, default "Times New Roman".}

\item{filesavename}{File savename.}

\item{savedir}{Save outputs to a temporary directory (default) else. Change to current directory e.g. "/home/me/folder". Do not use getwd() here. No terminal slash.}

\item{receiverlats}{Vector of latitudes for receivers to be plotted.}

\item{receiverlons}{Vector of longitudes for receivers to be plotted.}

\item{receivernames}{Vector of names for receivers to be plotted.}

\item{receiverrange}{Single (will be recycled), or vector of detection ranges in metres for receivers to be plotted.}

\item{recpointscol}{Colour of receiver centrepoint outlines.}

\item{recpointsfill}{Colour of receiver centrepoint fills.}

\item{recpointsalpha}{Alpha value of receiver centrepoint fills, 0 (invisible) to 1 (fully visible).}

\item{recpointssize}{Size of receiver points.}

\item{recpointsshape}{Shape of receiver points, default 21, circle with outline and fill.}

\item{recbufcol}{Colour of the receiver buffer circle outlines.}

\item{recbuffill}{Colour of the receiver buffer circle fills.}

\item{recbufalpha}{Alpha value of receiver buffer fills, 0 (invisible) to 1 (fully visible).}

\item{reclabcol}{Receiver label text colour.}

\item{reclabfill}{Receiver label fill colour, NA for no fill.}

\item{reclabnudgex}{Receiver label offset nudge in X dimension.}

\item{reclabnudgey}{Receiver label offset nudge in Y dimension.}

\item{reclabpad}{Receiver label padding in lines.}

\item{reclabrad}{Receiver label radius in lines.}

\item{reclabbord}{Receiver label border in mm.}

\item{surface}{Plot complete UD surface along with contours.}
}
\description{
Plots a group-level utilization distribution on a map background
}
\details{
Description paragraph: Automates delta log normal boosted regression trees abundance prediction.
Loops through all permutations of parameters provided (learning
rate, tree complexity, bag fraction), chooses the best, then simplifies it.
Generates line, dot and bar plots, and outputs these and the predictions
and a report of all variables used, statistics for tests, variable
interactions, predictors used and dropped, etc. If selected, generates
predicted abundance maps, and Unrepresentativeness surfaces.
See www.GitHub.com/SimonDedman/gbm.auto for issues, feedback, and development
suggestions. See SimonDedman.com for links to walkthrough paper, and papers
and thesis published using this package.
}
\author{
Simon Dedman, \email{simondedman@gmail.com}

Maurits van Zinnicq Bergmann, \email{mauritsvzb@gmail.com}
}
