% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleraster.R
\name{scaleraster}
\alias{scaleraster}
\title{Automated Boosted Regression Tree modelling and mapping suite}
\usage{
scaleraster(
  path = NULL,
  pattern = ".asc",
  format = "ascii",
  datatype = "FLT4S",
  bylayer = TRUE,
  overwrite = TRUE,
  scalefolder = "Scaled",
  summedname = "All_Rasters_Summed",
  scaledname = "All_Rasters_Scaled",
  crsloc = NULL,
  returnObj = FALSE
)
}
\arguments{
\item{path}{No terminal slash.}

\item{pattern}{Default ".asc".}

\item{format}{Default "ascii".}

\item{datatype}{Default "FLT4S".}

\item{bylayer}{Default TRUE.}

\item{overwrite}{Default TRUE.}

\item{scalefolder}{Default "Scaled".}

\item{summedname}{Default "All_Rasters_Summed".}

\item{scaledname}{Default "All_Rasters_Scaled".}

\item{crsloc}{Location of saved CRS Rds file from dBBMM.build.R. Should be same as path.}

\item{returnObj}{Logical. Return the scaled object to the parent environment? Default FALSE.}
}
\value{
Line, dot and bar plots, a report of all variables used, statistics
for tests, variable interactions, predictors used and dropped, etc. If
selected generates predicted abundance maps, and Unrepresentativeness surface
}
\description{
Automates delta log normal boosted regression trees abundance prediction.
Loops through all permutations of parameters provided (learning
rate, tree complexity, bag fraction), chooses the best, then simplifies it.
Generates line, dot and bar plots, and outputs these and the predictions
and a report of all variables used, statistics for tests, variable
interactions, predictors used and dropped, etc. If selected, generates
predicted abundance maps, and Unrepresentativeness surfaces.
See www.GitHub.com/SimonDedman/gbm.auto for issues, feedback, and development
suggestions. See SimonDedman.com for links to walkthrough paper, and papers
and thesis published using this package.
}
\details{
Errors and their origins:
}
\examples{
\donttest{
# Not run
}

}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
