% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TracksCleaned}
\alias{TracksCleaned}
\title{Data: Tracks of lemon sharks off Bimini, Bahamas}
\format{
A data frame with 1308 rows and 5 variables:
\describe{
  \item{Datetime}{POSIXct datetime, format YYYY-MM-DD HH:MM:SS}
  \item{Shark}{Individual shark ID code}
  \item{T.Ph}{Tidal phase}
  \item{Lat}{Decimal latitudes}
  \item{Lon}{Decimal longitudes}
}
}
\source{
\url{http://datras.ices.dk}
}
\usage{
data(TracksCleaned)
}
\description{
1 paragraph description
}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
\keyword{datasets}
