% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TracksCleaned}
\alias{TracksCleaned}
\title{Data: Tracks of lemon sharks off Bimini, Bahamas}
\format{
A data frame with 1308 rows and 5 variables:
\describe{
  \item{Datetime}{POSIXct datetime, format YYYY-MM-DD HH:MM:SS.}
  \item{Shark}{Individual shark ID code.}
  \item{T.Ph}{Tidal phase, H M L High Medium Low.}
  \item{Lat}{Decimal latitudes.}
  \item{Lon}{Decimal longitudes.}
}
}
\source{
\url{https://www.biminisharklab.com}
}
\usage{
data(TracksCleaned)
}
\description{
Tracks of 17 lemon sharks (Negaprion brevirostris) tagged off Bimini, Bahamas, 2012:2014, by 
Bimini Biological Field Station employees and volunteers, with accompanying tidal phase.
}
\author{
Simon Dedman, \email{simondedman@gmail.com}

Maurits van Zinnicq Bergmann, \email{mauritsvzb@gmail.com}
}
\keyword{datasets}
