% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightraster.R
\name{weightraster}
\alias{weightraster}
\title{Succinct title 8 words max}
\usage{
weightraster(
  locations =
    c("/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/H/Scaled",
    "/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/L/Scaled",
    "/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/M/Scaled"),
  rasternames = c("H", "L", "M"),
  pattern = "All_Rasters_Summed.asc",
  weightings = c(1, 2, 3),
  saveloc =
    "/home/simon/Dropbox/PostDoc Work/Rob Bullock accelerometer Lemons 2020.09/dBBMM ASCII/Weighted",
  extension = ".asc",
  format = "ascii",
  datatype = "FLT4S",
  bylayer = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{locations}{Character vector of locations of rasters to weight, e.g. locations of All_Rasters_Summed.asc. No terminal slash.}

\item{rasternames}{Same length character vector of names for rasters from locations.}

\item{pattern}{Names of the raster files in locations, e.g. "All_Rasters_Summed.asc".}

\item{weightings}{Same length numeric vector of weightings to apply to rasters. Rasters from locations are divided by their respective value from weightings e.g. raster 1 values divided by weightings 1 value.}

\item{saveloc}{Where to save output rasters. No terminal slash.}

\item{extension}{Extension format of output rasters.}

\item{format}{Raster format for output rasters.}

\item{datatype}{Data type for output rasters.}

\item{bylayer}{Raster saving detail.}

\item{overwrite}{Raster saving detail.}
}
\description{
Description paragraph: Automates delta log normal boosted regression trees abundance prediction.
Loops through all permutations of parameters provided (learning
rate, tree complexity, bag fraction), chooses the best, then simplifies it.
Generates line, dot and bar plots, and outputs these and the predictions
and a report of all variables used, statistics for tests, variable
interactions, predictors used and dropped, etc. If selected, generates
predicted abundance maps, and Unrepresentativeness surfaces.
See www.GitHub.com/SimonDedman/gbm.auto for issues, feedback, and development
suggestions. See SimonDedman.com for links to walkthrough paper, and papers
and thesis published using this package.
}
